requirements.txt
numpy
torch
torchvision
scikit-learn
tqdm
pandas
visdom
pillow
facenet-pytorch
flask
bootstrap-flask
python-dotenv